Update openzwave-db
I ran into the issue that my Fibaro Dimmer 2 (FGD-212) wasn't recognized and resulted in the following product in the control panel:

Unknown: type=0102, id=1000 (Unknown: id=010f)

Looking at the commits done on the OpenZwave repo after the previous version that was downloaded I expect this dimmer to be recognized. I am not sure why `3d6374c831998595744cc34ef862a80ee51973c9` was used until now, but this PR updates to the latest on `master`. Is there a reason not to update to the latest OpenZwave DB?